adding another Edelweiss use-case
I wrote some views/lists that help find docs that match certain criteria. These criteria are just data type (calibration, WIMP search) and temperature (which are the first-level cuts when trying to find interesting data). However, this might be an example of what NOT to do... if I were to do it over (which I may do), I would use search index.